+++ external_push/views.py
@@ -7,7 +7,7 @@ from django.utils import timezone
 from django.http import JsonResponse, HttpResponse
 from django.core.exceptions import ObjectDoesNotExist
 
-from .models import GenericPushTarget
+from .models import GenericPushTarget, BrewersFriendPushTarget
 
 import fermentrack_django.settings as settings
 
@@ -22,16 +22,23 @@ logger = logging.getLogger(__name__)
 
 @login_required
 @site_is_configured
-def external_push_list(request):
+def external_push_list(request, context_only=False):
     # TODO - Add user permissioning
     # if not request.user.has_perm('app.add_device'):
     #     messages.error(request, 'Your account is not permissioned to add devices. Please contact an admin')
     #     return redirect("/")
 
     all_push_targets = GenericPushTarget.objects.all()
+    brewers_friend_push_targets = BrewersFriendPushTarget.objects.all()
 
-    return render(request, template_name='external_push/push_target_list.html',
-                  context={'all_push_targets': all_push_targets})
+    context = {'all_push_targets': all_push_targets,
+               'brewers_friend_push_targets': brewers_friend_push_targets}
+
+    # This allows us to embed this in the site configuration page... There's almost certainly a better way to do this.
+    if not context_only:
+        return render(request, template_name='external_push/push_target_list.html', context=context)
+    else:
+        return context
 
 
 
@@ -109,6 +116,82 @@ def external_push_delete(request, push_target_id):
     messages.success(request, message)
 
     return redirect('external_push_list')
+
+
+
+
+@login_required
+@site_is_configured
+def external_push_brewers_friend_target_add(request):
+    # TODO - Add user permissioning
+    # if not request.user.has_perm('app.add_device'):
+    #     messages.error(request, 'Your account is not permissioned to add devices. Please contact an admin')
+    #     return redirect("/")
+
+    form = forms.BrewersFriendPushTargetModelForm()
+
+    if request.POST:
+        form = forms.BrewersFriendPushTargetModelForm(request.POST)
+        if form.is_valid():
+            new_push_target = form.save()
+            messages.success(request, 'Successfully added push target')
+            return redirect('external_push_list')
+
+        messages.error(request, 'Unable to add new push target')
+
+    # Basically, if we don't get redirected, in every case we're just outputting the same template
+    return render(request, template_name='external_push/brewers_friend_push_target_add.html', context={'form': form})
+
+
+@login_required
+@site_is_configured
+def external_push_brewers_friend_view(request, push_target_id):
+    # TODO - Add user permissioning
+    # if not request.user.has_perm('app.add_device'):
+    #     messages.error(request, 'Your account is not permissioned to add devices. Please contact an admin')
+    #     return redirect("/")
+
+    try:
+        push_target = BrewersFriendPushTarget.objects.get(id=push_target_id)
+    except ObjectDoesNotExist:
+        messages.error(request, "Brewers's Friend push target {} does not exist".format(push_target_id))
+        return redirect('external_push_list')
+
+    if request.POST:
+        form = forms.BrewersFriendPushTargetModelForm(request.POST, instance=push_target)
+        if form.is_valid():
+            updated_push_target = form.save()
+            messages.success(request, 'Updated push target')
+            return redirect('external_push_list')
+
+        messages.error(request, 'Unable to update push target')
+
+    form = forms.BrewersFriendPushTargetModelForm(instance=push_target)
+
+    return render(request, template_name='external_push/brewers_friend_push_target_view.html',
+                  context={'push_target': push_target, 'form': form})
+
+
+@login_required
+@site_is_configured
+def external_push_brewers_friend_delete(request, push_target_id):
+    # TODO - Add user permissioning
+    # if not request.user.has_perm('app.add_device'):
+    #     messages.error(request, 'Your account is not permissioned to add devices. Please contact an admin')
+    #     return redirect("/")
+
+    try:
+        push_target = BrewersFriendPushTarget.objects.get(id=push_target_id)
+    except ObjectDoesNotExist:
+        messages.error(request, "Brewers's Friend push target {} does not exist".format(push_target_id))
+        return redirect('external_push_list')
+
+    message = "Brewers's Friend push target {} has been deleted".format(push_target_id)
+    push_target.delete()
+    messages.success(request, message)
+
+    return redirect('external_push_list')
+